requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.0
pymongo==4.13.0
pydantic==2.9.2
motor==3.7.1
python-dotenv
requests==2.31.0
pytest
pytest-asyncio
httpx
Jinja2>=3.1.2
dnspython
python-dotenv